refactor(library): extract seeded-radio builder into a service
